test_projects/anthropic_claude_3_7_sonnet_python/requirements.txt

Flask==2.0.1
Werkzeug==2.0.1
psycopg2-binary==2.9.1
